rename TokenIcon to AssetIcon
