+++ src/components/TokenTitle.tsx
@@ -2,7 +2,7 @@ import { AssetInfo } from '@ergolabs/ergo-sdk/build/main/entities/assetInfo';
 import React, { FC } from 'react';
 
 import { Flex, Typography } from '../ergodex-cdk';
-import { TokenIcon } from './TokenIcon/TokenIcon';
+import { AssetIcon } from './AssetIcon/AssetIcon';
 
 export interface TokenTitleProps {
   readonly asset: AssetInfo;
@@ -12,7 +12,7 @@ export interface TokenTitleProps {
 export const TokenTitle: FC<TokenTitleProps> = ({ asset, size }) => (
   <Flex align="center">
     <Flex.Item marginRight={1}>
-      <TokenIcon size={size} asset={asset} />
+      <AssetIcon size={size} asset={asset} />
     </Flex.Item>
     <Typography.Title level={5}>{asset.name}</Typography.Title>
   </Flex>
