+++ src/components/WalletModal/TokensTab/TokenListItem/TokenListItem.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Currency } from '../../../../common/models/Currency';
 import { Box, Flex, Typography } from '../../../../ergodex-cdk';
-import { TokenIcon } from '../../../TokenIcon/TokenIcon';
+import { AssetIcon } from '../../../AssetIcon/AssetIcon';
 
 interface TokenListItemProps {
   readonly currency: Currency;
@@ -14,7 +14,7 @@ export const TokenListItem: React.FC<TokenListItemProps> = ({ currency }) => (
       <Flex.Item flex={1}>
         <Flex align="center">
           <Flex.Item marginRight={2}>
-            <TokenIcon asset={currency.asset} />
+            <AssetIcon asset={currency.asset} />
           </Flex.Item>
           <Flex direction="col">
             <Typography.Body>{currency.asset.name}</Typography.Body>
