+++ src/components/WalletModal/WalletTotalBalance/WalletTotalBalance.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { Currency } from '../../../common/models/Currency';
 import { Box, Flex, LoadingOutlined, Typography } from '../../../ergodex-cdk';
-import { TokenIcon } from '../../TokenIcon/TokenIcon';
+import { AssetIcon } from '../../AssetIcon/AssetIcon';
 
 interface WalletTotalBalanceProps {
   balance?: Currency;
@@ -22,7 +22,7 @@ export const WalletTotalBalance: React.FC<WalletTotalBalanceProps> = ({
       <Box padding={[2, 4]} borderRadius="m" contrast>
         <Flex row align="center">
           <Flex.Item marginRight={2}>
-            <TokenIcon asset={balance?.asset} />
+            <AssetIcon asset={balance?.asset} />
           </Flex.Item>
           <Flex.Item flex={1}>
             <Typography.Title level={4}>
