+++ src/components/common/TokenControl/TokenSelect/TokenListModal/TokenListItem.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import { Box, Flex, Typography } from '../../../../../ergodex-cdk';
 import { useAssetsBalance } from '../../../../../gateway/api/assetBalance';
-import { TokenIcon } from '../../../../TokenIcon/TokenIcon';
+import { AssetIcon } from '../../../../AssetIcon/AssetIcon';
 import { Truncate } from '../../../../Truncate/Truncate';
 
 interface TokenListItemProps {
@@ -26,7 +26,7 @@ const TokenListItem: React.FC<TokenListItemProps> = ({ asset, onClick }) => {
       bordered={false}
     >
       <Flex align="center" style={{ width: '100%' }}>
-        <TokenIcon size="large" asset={asset} />
+        <AssetIcon size="large" asset={asset} />
         <Box className="token-item__box" padding={0} bordered={false}>
           <Box className="token-item__box-left" padding={0} bordered={false}>
             <Typography.Text className="token-item__box-left-symbol">
