+++ src/pages/Pool/components/LiquidityPositionsList/LiquidityPositionsItem/LiquidityPositionsItemWrapper/LiquidityPositionsItemWrapper.tsx
@@ -3,9 +3,9 @@ import { Trans } from '@lingui/macro';
 import React from 'react';
 
 import { AmmPool } from '../../../../../../common/models/AmmPool';
+import { TokenIconPair } from '../../../../../../components/AssetIconPair/TokenIconPair';
 import { DataTag } from '../../../../../../components/common/DataTag/DataTag';
 import { ListItemWrapper } from '../../../../../../components/ListItemWrapper/ListItemWrapper';
-import { TokenIconPair } from '../../../../../../components/TokenIconPair/TokenIconPair';
 import { Truncate } from '../../../../../../components/Truncate/Truncate';
 import { VerificationMark } from '../../../../../../components/VerificationMark/VerificationMark';
 import { Flex, Typography } from '../../../../../../ergodex-cdk';
