+++ src/pages/Swap/PoolSelector/PoolSelector.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components';
 
 import { useSubject } from '../../../common/hooks/useObservable';
 import { AmmPool } from '../../../common/models/AmmPool';
+import { TokenIconPair } from '../../../components/AssetIconPair/TokenIconPair';
 import { DataTag } from '../../../components/common/DataTag/DataTag';
 import { InfoTooltip } from '../../../components/InfoTooltip/InfoTooltip';
-import { TokenIconPair } from '../../../components/TokenIconPair/TokenIconPair';
 import { Truncate } from '../../../components/Truncate/Truncate';
 import { VerificationMark } from '../../../components/VerificationMark/VerificationMark';
 import {
